# gin-blog-server/docs/swagger.yaml
definitions:
...
  handler.AddOrEditTagReq:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  handler.PageResult-model_RoleVO:
    properties:
      page_data:
        description: 分页数据
        items:
          $ref: '#/definitions/model.RoleVO'
        type: array
      page_num:
        description: 每页条数
        type: integer
      page_size:
        description: 上次页数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  handler.PageResult-model_TagVO:
    properties:
      page_data:
        description: 分页数据
        items:
          $ref: '#/definitions/model.TagVO'
        type: array
      page_num:
        description: 每页条数
        type: integer
      page_size:
        description: 上次页数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
...
  handler.Response-handler_PageResult-model_RoleVO:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/handler.PageResult-model_RoleVO'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-handler_PageResult-model_TagVO:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/handler.PageResult-model_TagVO'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-int:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
        type: integer
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.Response-model_OptionVO:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.OptionVO'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.Response-model_Tag:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.Tag'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  model.Article:
    properties:
      category_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      desc:
        type: string
      id:
        type: integer
      img:
        type: string
      is_delete:
        type: integer
      is_top:
        type: integer
      original_url:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  model.OptionVO:
    properties:
      label:
        type: string
      value:
        type: integer
    type: object
...
  model.RoleVO:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_disable:
        type: boolean
      label:
        type: string
      menu_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      resource_ids:
        items:
          type: integer
        type: array
    type: object
  model.Tag:
    properties:
      articles:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.TagVO:
    properties:
      article_count:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /role/list:
    get:
      description: 获取角色列表
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-handler_PageResult-model_RoleVO'
      summary: 获取角色列表
      tags:
      - role
...
  /tag:
    delete:
      consumes:
      - application/json
      description: 根据 ID 数组删除标签
      parameters:
      - description: 标签 ID 数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int'
      security:
      - ApiKeyAuth: []
      summary: 删除标签（批量）
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: 添加或修改标签
      parameters:
      - description: 添加或修改标签
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.AddOrEditTagReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_Tag'
      security:
      - ApiKeyAuth: []
      summary: 添加或修改标签
      tags:
      - Tag
  /tag/list:
    get:
      consumes:
      - application/json
      description: 根据条件查询获取标签列表
      parameters:
      - description: 当前页数
        in: query
        name: page_size
        type: integer
      - description: 每页条数
        in: query
        name: page_num
        type: integer
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/handler.Response-handler_PageResult-model_TagVO'
      security:
      - ApiKeyAuth: []
      summary: 获取标签列表
      tags:
      - Tag
  /tag/option:
    get:
      consumes:
      - application/json
      description: 获取标签选项列表
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_OptionVO'
      security:
      - ApiKeyAuth: []
      summary: 获取标签选项列表
      tags:
      - Tag
...
swagger: "2.0"